fix remove qdrant semantic caching to it's own folder
